Add reusable inject extra information function
